app-ai-simple/pyproject.toml
[project]
name = "app-ai-simple"
version = "1.0.0"
description = "LangChain ReAct AI agent powered by Claude"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.135",
    "uvicorn[standard]>=0.41",
    "python-multipart>=0.0.22",

    # Validation
    "pydantic>=2.12",
    "pydantic-settings>=2.13",

    # LangChain
    "langchain-core>=1.2",
    "langchain-anthropic>=1.3",
    "langchain-text-splitters>=1.1",
    "langchain-tavily>=0.2",

    # Vector DB
    "pymilvus>=2.6",
    "milvus-lite>=2.5",

    # Embeddings (BGE-M3, local model)
    "FlagEmbedding>=1.3",
    "modelscope>=1.34",
    "sentence-transformers>=5.2",
    "torch>=2.10",
    "transformers==4.44.2",
    "accelerate>=1.13",

    # Document parsing
    "pymupdf>=1.27",
    "pytesseract",
    "pillow",
    "python-docx",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
